2.0.5 (#184)
* refactoring

* refactoring

refactoring

refactoring & #46

Create help section for configuration and #47 #61

changelog

[tx-robot] updated from transifex


[tx-robot] updated from transifex


[tx-robot] updated from transifex


[tx-robot] updated from transifex


* transifex

* transifex

* test

* Translation

* translation

* hotfix

* https://github.com/baimard/gestion/issues/123

* Update lib/Migration/Version20002Date20220201085856.php


* Update lib/Migration/Version20002Date20220201085856.php


* Update lib/Migration/Version7Date20220125174818.php


* Update lib/Migration/Version7Date20220125174818.php


* Update lib/Migration/Version7Date20220125174818.php


* https://github.com/baimard/gestion/issues/135

* Update lib/Migration/Version20002Date20220201085856.php


* valide

* valide

* final

* Final

* Final

* Update lib/Migration/Version20002Date20220201085856.php


* Update lib/Migration/Version20002Date20220201085856.php


* Update lib/Migration/Version7Date20220125174818.php


* Update lib/Migration/Version7Date20220125174818.php


* Update lib/Migration/Version7Date20220125174818.php


* [tx-robot] updated from transifex


* l10n: Fixed grammar

Reported at Transifex.


* l10n: Improve styling

A more selected description of the function setting.

* l10n: Improve styling 

A more selected description of the function setting.

* l10n: Add "an"

* l10n: Add "an"

* l10n: Replace a word

* l10n: Replace a word

* [tx-robot] updated from transifex


* Update lib/Migration/Version20002Date20220201085856.php


* valide

* valide

* final

* Final

* Final

* Hotfix 2.0.2

* Hotfix 2.0.3

* enhancement tests !

* https://github.com/baimard/gestion/issues/143

* Postgresql

* NPM UPDATE

* https://github.com/baimard/gestion/issues/168

* https://github.com/baimard/gestion/issues/167

* https://github.com/baimard/gestion/issues/163

* https://github.com/baimard/gestion/issues/158

* https://github.com/baimard/gestion/issues/160

* https://github.com/baimard/gestion/issues/173

* Bug in configuration and delete confirmation message

+++ src/js/modules/ajaxRequest.mjs
@@ -26,13 +26,13 @@ export function updateDB(table, column, data, id) {
     }).done(function (response, code) {
         showSuccess(t('gestion', 'Modification saved'));
     }).fail(function (response, code) {
-        showError(response);
+        showError(t('gestion', 'There is an error with the format, please check the documentation'));
     });
 }
 
 /**
  * Delete data
- * @param table 
+ * @param table
  * @param id 
  */
 export function deleteDB(table, id) {
@@ -41,17 +41,21 @@ export function deleteDB(table, id) {
         id: id,
     };
 
-    $.ajax({
-        url: baseUrl + '/delete',
-        type: 'DELETE',
-        async: false,
-        contentType: 'application/json',
-        data: JSON.stringify(myData)
-    }).done(function (response, code) {
-        showMessage(t('gestion', 'Modification saved'));
-    }).fail(function (response, code) {
-        showError(response);
-    });
+    if(window.confirm(t('gestion','Are you sure you want to delete?'))){
+        $.ajax({
+            url: baseUrl + '/delete',
+            type: 'DELETE',
+            async: false,
+            contentType: 'application/json',
+            data: JSON.stringify(myData)
+        }).done(function (response, code) {
+            showSuccess(t('gestion', 'Modification saved'));
+        }).fail(function (response, code) {
+            showError(response);
+        });
+    }else{
+        showMessage(t('gestion', 'Nothing changed'))
+    }
 }
 
 /**
@@ -133,9 +137,12 @@ export function getAnnualTurnoverPerMonthNoVat(cur) {
                 curRow = insertRow("Statistical", -1, 0, item.y);
                 modifyCell(curRow, (item.m), cur.format(Math.round(item.total)));
                 total+= Math.round(item.total);
+
             }else{
+
                 modifyCell(curRow, (item.m), cur.format(Math.round(item.total)));
                 total+= Math.round(item.total);
+
             }
         });
         // At the end
@@ -170,7 +177,7 @@ export function listProduit(lp, id, produitid) {
         type: 'PROPFIND',
         contentType: 'application/json'
     }).done(function (response) {
-        lp.append('<option data-table="produit_devis" data-column="produit_id" data-val="' + produitid + '" data-id="' + id + '">Annuler</option>');
+        lp.append('<option data-table="produit_devis" data-column="produit_id" data-val="' + produitid + '" data-id="' + id + '">'+t('gestion','Cancel')+'</option>');
         $.each(JSON.parse(response), function (arrayID, myresp) {
             var selected = "";
             if (produitid == myresp.id) {
